XAI Application/requirements.txt
pip==25.1.1
python-dotenv
requests
pydantic
fastapi